semigrp+elements: Adds transformation membership
Adds membership testing function for transformations of
threshold 1 in commutative transformation semigroups.

Adds membership testing method for permuation groups. This is used
in the above method.

Implements both of the above methods to be called automatically in
the Semigroups class __contains__ method.

Includes extra functions used for unittesting: SymmetricGroup,
TransDihedralGroup and TransDirectProduct and the member methods
for the Semigroup class, IsomorphicTransSemigroup and copy().

Also adds extra functionality for the Transformation class,
including functions to convert a permutation into disjoint cycle
notation, and to find a permutation's inverse. These are called in
the membership functions.

Includes unittests and documentation.